requirements.txt
absl-py==2.1.0
actionlib==1.14.0
angles==1.9.13
ansitable==0.9.10
antlr4-python3-runtime==4.8
arm_pytorch_utilities==0.4.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
base_local_planner==1.17.3
cachetools==4.2.4
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1707022139797/work/certifi
cfgv==3.3.1
charset-normalizer==3.3.2
cloudpickle==2.2.1
colored==1.4.4
comm==0.1.4
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cv-bridge==1.16.2
cycler==0.11.0
Cython==3.0.9
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
dgl==0.9.1
diagnostic-updater==1.11.0
distlib==0.3.8
dynamic-reconfigure==1.7.3
empy==4.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
exceptiongroup==1.2.0
filelock==3.12.2
fonttools==4.38.0
freetype-py==2.4.0
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
glfw==2.7.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
graphviz==0.20.1
grpcio==1.62.1
gym==0.20.0
h5py==3.8.0
hydra-core==1.1.1
identify==2.5.24
idna==3.6
ikpy==3.3.4
image-geometry==1.16.2
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
interactive-markers==1.12.0
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1620912939357/work/dist/ipykernel-5.5.5-py3-none-any.whl
ipympl==0.9.3
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
joblib==1.3.2
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
jupyter-client==5.3.4
jupyter_core==4.12.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
laser_geometry==1.6.7
-e git+https://github.com/iROSA-lab/rlmmbp.git@341cf0c24b64aa0329cd4b217f8fbc0d080a584a#egg=learned_robot_placement
libxsmm==0.0.1
lxml==5.2.2
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
meshcat==0.3.2
message-filters==1.16.0
moveit-core==1.1.13
moviepy==1.0.3
mpmath==1.3.0
mujoco==2.3.6
-e git+https://github.com/iROSA-lab/mushroom-rl.git@20e6ca13f367c9f751dbb0c593a05b80d00b5fe7#egg=mushroom_rl
networkx==2.2
nodeenv==1.8.0
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1649806299270/work
numpy-ml==0.1.2
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.9.0.80
packaging==24.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1706113125309/work
pgraph-python==0.6.2
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs==4.0.0
pluggy==1.2.0
pre-commit==2.21.0
proglog==0.1.10
progress==1.6
prompt-toolkit==3.0.43
protobuf==3.20.1
psutil==5.9.8
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.12.0
pycollada==0.6
pygame==2.5.2
pyglet==2.0.10
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1700607939962/work
pyngrok==7.0.5
PyOpenGL==3.1.0
pyparsing==3.1.2
pyrender==0.1.45
pytest==7.4.4
python-dateutil==2.9.0.post0
python-qt-binding==0.4.4
-e git+https://github.com/iROSA-lab/sampled_reachability_maps.git@cd49112daebac83e4f66ab145b8c2aa78a1477d9#egg=pytorch_kinematics
pytorch-seed==0.2.0
PyYAML==6.0.1
pyzmq==25.1.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
requests==2.31.0
requests-oauthlib==1.4.0
resource_retriever==1.12.7
roboticstoolbox-python==1.1.0
rosbag==1.16.0
rosclean==1.15.8
rosgraph==1.16.0
roslaunch==1.16.0
roslib==1.15.8
roslz4==1.16.0
rosmaster==1.16.0
rosmsg==1.16.0
rosnode==1.16.0
rosparam==1.16.0
rospy==1.16.0
rosserial_arduino==0.9.2
rosserial_client==0.9.2
rosserial_python==0.9.2
rosservice==1.16.0
rostest==1.16.0
rostopic==1.16.0
rosunit==1.15.8
roswtf==1.16.0
rqt-image-view==0.4.17
rqt_gui==0.5.3
rsa==4.9
rtb-data==1.0.1
scikit-learn==1.0.2
scipy==1.7.3
sensor-msgs==1.13.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smclib==1.8.6
spatialgeometry==1.1.0
spatialmath-python==1.1.9
srdfdom==0.6.4
swift-sim==1.1.0
sympy==1.10.1
tensorboard==2.4.0
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
threadpoolctl==3.1.0
tomli==2.0.1
topic-tools==1.16.0
torch==1.12.0+cu113
torchaudio==0.11.0+cu113
torchvision==0.13.0+cu113
tornado==6.2
tqdm==4.64.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1675110562325/work
transformations==2021.6.6
trimesh==4.4.0
typing_extensions==4.7.1
u-msgpack-python==2.8.0
urdfdom-py==0.4.6
urdfpy==0.0.22
urllib3==2.0.7
virtualenv==20.25.1
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==2.2.3
widgetsnbextension==4.0.10
xacro==1.14.16
zipp==3.15.0
